[rocketchip] split out Base and Example tops
